tests/requirements-test.txt
# Test dependencies for Bitcoin data pipeline
pytest>=7.0.0
pytest-cov>=4.0.0
boto3>=1.26.0
botocore>=1.29.0
aws-cdk-lib>=2.0.0
aws-cdk.assertions>=1.200.0
